Refactor templateBoxPlot
